requirements.txt
apiflask==2.1.1
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
cryptography==42.0.7
random-username==1.0.2
urllib3==2.2.1
boto3==1.34.115